Move debug cen package to debug
