+++ app/src/main/java/org/coepi/android/ui/debug/DebugPagerAdapter.kt
@@ -2,7 +2,7 @@ package org.coepi.android.ui.debug
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import org.coepi.android.ui.cen.CENFragment
+import org.coepi.android.ui.debug.cen.CENFragment
 import org.coepi.android.ui.debug.logs.LogsFragment
 
 class DebugPagerAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
