+++ app/src/main/java/org/coepi/android/ui/debug/cen/CENRecyclerAdapter.kt
@@ -1,25 +1,25 @@
-package org.coepi.android.ui.cen
+package org.coepi.android.ui.debug.cen
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import org.coepi.android.R.layout.item_cen_report
-import org.coepi.android.cen.RealmCenReport
-import org.coepi.android.ui.cen.CENReportViewAdapter.ViewHolder
+import org.coepi.android.R.layout.item_cen
+import org.coepi.android.ui.debug.cen.CENRecyclerViewAdapter.ViewHolder
 
-class CENReportViewAdapter : RecyclerView.Adapter<ViewHolder>() {
-    private var items = emptyList<RealmCenReport>()
+class CENRecyclerViewAdapter : RecyclerView.Adapter<ViewHolder>() {
+    private var items = emptyList<String>()
 
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         private val textView = view as TextView
-        fun bind(item: RealmCenReport) {
-            textView.text = item.report
+        fun bind(item: String) {
+            textView.text = item
         }
+
     }
 
-    fun setItems(items: List<RealmCenReport>) {
+    fun setItems(items: List<String>) {
         this.items = items
         notifyDataSetChanged()
     }
@@ -27,7 +27,7 @@ class CENReportViewAdapter : RecyclerView.Adapter<ViewHolder>() {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
         ViewHolder(
             LayoutInflater.from(parent.context).inflate(
-                item_cen_report, parent,
+                item_cen, parent,
                 false
             )
         )
