+++ app/src/main/java/org/coepi/android/ui/debug/cen/CENReportViewAdapter.kt
@@ -1,25 +1,25 @@
-package org.coepi.android.ui.cen
+package org.coepi.android.ui.debug.cen
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import org.coepi.android.R.layout.item_cen
-import org.coepi.android.ui.cen.CENRecyclerViewAdapter.ViewHolder
+import org.coepi.android.R.layout.item_cen_report
+import org.coepi.android.cen.RealmCenReport
+import org.coepi.android.ui.debug.cen.CENReportViewAdapter.ViewHolder
 
-class CENRecyclerViewAdapter : RecyclerView.Adapter<ViewHolder>() {
-    private var items = emptyList<String>()
+class CENReportViewAdapter : RecyclerView.Adapter<ViewHolder>() {
+    private var items = emptyList<RealmCenReport>()
 
     class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         private val textView = view as TextView
-        fun bind(item: String) {
-            textView.text = item
+        fun bind(item: RealmCenReport) {
+            textView.text = item.report
         }
-
     }
 
-    fun setItems(items: List<String>) {
+    fun setItems(items: List<RealmCenReport>) {
         this.items = items
         notifyDataSetChanged()
     }
@@ -27,7 +27,7 @@ class CENRecyclerViewAdapter : RecyclerView.Adapter<ViewHolder>() {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
         ViewHolder(
             LayoutInflater.from(parent.context).inflate(
-                item_cen, parent,
+                item_cen_report, parent,
                 false
             )
         )
